requirements.txt
# Versions compatibles pour LangChain avec Ollama
langchain>=0.1.12
langchain-core>=0.2.20,<0.3.0
langchain-community>=0.0.29
langchain-ollama>=0.1.0

# LangChain dependencies
pydantic>=2.6.4
pydantic-core>=2.16.3

# Embeddings
gpt4all>=2.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Recherche de similarité améliorée
faiss-cpu>=1.7.4  # Pour environnements CPU - principal backend pour la recherche vectorielle
# annoy>=1.17.0   # Alternative légère à FAISS (nécessite un compilateur C++)
cross-encoders>=0.0.4  # Pour le reranking avancé

# Image processing
pillow>=10.0.0
matplotlib>=3.7.0
requests>=2.31.0

# Multimodal RAG avec transformers
torch>=2.0.0
transformers>=4.30.0  # Inclut l'implémentation de CLIP
pymupdf>=1.22.0       # PyMuPDF pour le traitement des PDF
numpy>=1.24.0
tqdm>=4.65.0

# Utilitaires
ipykernel
pandas
jupyter

# Pour éviter les avertissements de dépendances
orjson>=3.9.10
tenacity>=8.2.3
typing-extensions>=4.9.0 